backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
Pillow==9.4.0
Pandas==1.5.3
djangorestframework==3.14.0
